一些类的整合

+++ app/src/main/java/com/ywl01/jlinfo/beans/PeopleBean.java
@@ -6,7 +6,7 @@ package com.ywl01.jlinfo.beans;
 
 public class PeopleBean{
 
-    public long id;
+    public int id;
     public String peopleNumber;
     public String name;
     public String namePinyin;
